Converted the explorability algorithm back to "permission trumps Site".
Now, if a User has permission to explore a Page, regardless of which Site the
Page is on, that User can explore it. Only if the User doesn't have permission
to explore the Page does the Page's associated Site come into play, to determine
whether a 404 or 403 error should be raised.