Massive refactor of the RequestCache mechanism, which necessitated changes to almost all the other new code.
The RequestCache is now stored as an attribute on the request objest itself
(because DUH... why didn't I think of that?). This allows the models.py to avoid
the need to import anything in the request/response layer. However, it also
complicates the permissions checking, because we must now pass the request all
the way down to the low perm checking functions, rather than the User. This
required some rather... creative changes, but I think everything worked out well
in the end.

This will be the final commit on this PR. I will be breaking up the changes in
this branch into multiple separate PRs, to make them more self-contained, and
easier to review.